requirements.txt
# Agentic AI Network Optimizer Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
openai>=1.3.0
pyyaml>=6.0
rich>=13.6.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
